Screen 1 collisions
Could not finish all the way, the drawingFunctions file is too large and over half the time I lost the address of where the blocks are. This has stopped me from getting to do the collisions for under the blocks. Still need help with jumping down from the blocks, and finding a solution for killing the shell enemy. I will be back in the lab tomorrow to try and fix these.