Add files via upload
Added screen transitions that can go both back and forth, I still need to make it so that if a block is destroyed it does not get re-created when going back to that screen. When you are testing the screen transitions may seem weird, but if you are encountering a bug where it resets, that's becuase you are still checking for collisions with the first turtle on the other screens. I also tweaked the jump a bit. I think the reason that he jump straight down is becuase we continually loop in jumpdown without allowing for controller input. Let me know if you get anything else done